// src/generated/scalar/protocol/v1beta1/query.d.ts
// @generated by protobuf-ts 2.9.4 with parameter long_type_string,output_javascript_es2015,generate_dependencies
// @generated from protobuf file "scalar/protocol/v1beta1/query.proto" (package "scalar.protocol.v1beta1", syntax proto3)
// tslint:disable
import type { BinaryWriteOptions } from "@protobuf-ts/runtime";
import type { IBinaryWriter } from "@protobuf-ts/runtime";
import type { BinaryReadOptions } from "@protobuf-ts/runtime";
import type { IBinaryReader } from "@protobuf-ts/runtime";
import type { PartialMessage } from "@protobuf-ts/runtime";
import { MessageType } from "@protobuf-ts/runtime";
import { Protocol } from "./types";
import { Status } from "./types";
/**
 * @generated from protobuf message scalar.protocol.v1beta1.ProtocolsRequest
 */
export interface ProtocolsRequest {
    /**
     * @generated from protobuf field: string pubkey = 1;
     */
    pubkey: string;
    /**
     * @generated from protobuf field: string address = 2;
     */
    address: string;
    /**
     * @generated from protobuf field: string name = 3;
     */
    name: string;
    /**
     * @generated from protobuf field: scalar.protocol.v1beta1.Status status = 4;
     */
    status: Status;
}
/**
 * @generated from protobuf message scalar.protocol.v1beta1.ProtocolsResponse
 */
export interface ProtocolsResponse {
    /**
     * @generated from protobuf field: repeated scalar.protocol.v1beta1.Protocol protocols = 1;
     */
    protocols: Protocol[];
}
declare class ProtocolsRequest$Type extends MessageType<ProtocolsRequest> {
    constructor();
    create(value?: PartialMessage<ProtocolsRequest>): ProtocolsRequest;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ProtocolsRequest): ProtocolsRequest;
    internalBinaryWrite(message: ProtocolsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message scalar.protocol.v1beta1.ProtocolsRequest
 */
export declare const ProtocolsRequest: ProtocolsRequest$Type;
declare class ProtocolsResponse$Type extends MessageType<ProtocolsResponse> {
    constructor();
    create(value?: PartialMessage<ProtocolsResponse>): ProtocolsResponse;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ProtocolsResponse): ProtocolsResponse;
    internalBinaryWrite(message: ProtocolsResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message scalar.protocol.v1beta1.ProtocolsResponse
 */
export declare const ProtocolsResponse: ProtocolsResponse$Type;
export {};